requirements.txt
alembic==1.7.7
blinker==1.4
click==8.1.3
config==0.5.1
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==1.0.2
Flask-Script==2.0.5
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.36
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
